Clarify the MessageBox trait / Deprecate the MessageBoxBuilder trait
The input/output contract of a message box if better expressed
at the level of the message box itself. The MessageBoxBuilder
is then less useful and has been deprecated.
However, one needs a place to build the message box for an actor.
This will be done by the ActorBuilder with no need for a specific abstraction.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>